added SplitEdgeAndCreateVertex() (tesselate) <== this was a brutal headache, eventually refactored to turn the method into bite-sized chunks, avoiding interdependencies, etc. etc. added SplitEdgeAndCreateVertex tesselation tests added CalculateEdgeCount around a vertex added drawing loop cycles to help debug them added ForEachRadialLoop enumerator added more utility methods to Edge and Loop updated the test Validation methods to account for more things that could possibly break refactoring here and there, as usual
NOTE: I decided to assume exactly 1 or 2 radial loops around an edge since three or more faces sharing the same edge is not a use case I've ever heard of for game engines (even if so, it's trivial to work around this on the modelling side and two-sided faces are either enabled in the shader or could be supported by mesh triangulation processing or duplicating faces or whatever). It's just a lot less headaches that way.